+++ inc/ti/do.h
@@ -19,7 +19,11 @@ int ti_do_ternary(ti_query_t * query, cleri_node_t * nd, ex_t * e);
 int ti_do_if_statement(ti_query_t * query, cleri_node_t * nd, ex_t * e);
 int ti_do_return_val(ti_query_t * query, cleri_node_t * nd, ex_t * e);
 int ti_do_return_alt_deep(ti_query_t * query, cleri_node_t * nd, ex_t * e);
+int ti_do_for_loop(ti_query_t * query, cleri_node_t * nd, ex_t * e);
+int ti_do_continue(ti_query_t * query, cleri_node_t * nd, ex_t * e);
+int ti_do_break(ti_query_t * query, cleri_node_t * nd, ex_t * e);
 int ti_do_closure(ti_query_t * query, cleri_node_t * nd, ex_t * e);
+int ti_do_prepare_for_loop(ti_query_t * query, cleri_node_t * vars_nd);
 int ti_do_init(void);
 void ti_do_drop(void);